Refactored flask_blog into a module
