+++ .gitignore
@@ -1,3 +1,4 @@
 .idea
 *__pycache__
 venv
+site.db
