+++ flask_blog/models/__init__.py
@@ -1,0 +1,2 @@
+from .user import User
+from .post import Post
