+++ flask_blog/models/user.py
@@ -1,0 +1,20 @@
+from flask_blog import db
+
+
+class User(db.Model):
+    __tablename__ = "user"
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(20), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    image_file = db.Column(db.String(20), nullable=False, default="default.jpeg")
+    password = db.Column(db.String(60), nullable=False)
+    posts = db.relationship('Post', backref='author', lazy=True)
+
+    def __repr__(self):
+        return f"User({self.username}, {self.email}, {self.image_file})"
+
+    def __init__(self, username, email, image_file, password):
+        self.username = username
+        self.email = email
+        self.image_file = image_file
+        self.password = password
