+++ flask_blog/routes/__init__.py
@@ -1,0 +1,5 @@
+from .about import about
+from .home import home
+from .login import login
+from .page_not_found import page_not_found
+from .register import register
