+++ flask_blog/routes/about.py
@@ -1,0 +1,7 @@
+from flask import render_template
+from flask_blog import app
+
+
+@app.route("/about")
+def about():
+    return render_template("about.html")
