+++ flask_blog/templates/register.html
