+++ requirements.txt
@@ -1,20 +1,29 @@
+alembic==1.7.7
 bcrypt==3.1.4
 blinker==1.4
 certifi==2016.2.28
-cffi
+cffi==1.15.0
 click==6.7
+colorama==0.4.4
 Flask==1.0
 Flask-Bcrypt==0.7.1
 Flask-Login==0.4.1
 Flask-Mail==0.9.1
-Flask-SQLAlchemy==2.3.2
+Flask-Migrate==3.1.0
+Flask-SQLAlchemy==2.5.1
 Flask-WTF==0.14.2
+greenlet==1.1.2
+importlib-metadata==4.11.3
+importlib-resources==5.6.0
 itsdangerous==0.24
 Jinja2==2.10
+Mako==1.2.0
 MarkupSafe==2.0.1
-Pillow
+Pillow==9.0.1
 pycparser==2.18
+pycryptodome==3.14.1
 six==1.11.0
-SQLAlchemy==1.2.7
+SQLAlchemy==1.4.34
 Werkzeug==0.14.1
 WTForms==2.1
+zipp==3.7.0
